requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
